requirements.txt
Django==1.6.5
Pillow==2.4.0
psycopg2==2.5.3
sorl-thumbnail==11.12
wsgiref==0.1.2
requests==2.3.0
